Added Windows Scripts and Checklist
Commenting is kinda minimal, easier to walk people through the functionality.
It is good practice to not rely on the scripts for points, so I have included a short writeup of a checklist that contains some instructions to enable and disable certain services that could score points within the competition. At this current time I am trying to access my old laptop which contains some more expansive checklists. 

Some of the functionality has not been implemented within WindowsFriendly.bat
Run all scripts as administrator from the desktop of the image.
Scripts have yet to be tested thoroughly and as such some functionality may not work as intended. Let me know if this is the case.
Run with caution.